Revert "Deltakere redesign"
