+++ monstring.class.php
@@ -593,20 +593,6 @@ class monstring_v2 {
 					$this->innslagTyper->addById( $r['bt_id'] );
 				}
 			}
-			if( $this->innslagTyper->getAntall() == 0 ) {
-				foreach( innslag_typer::getAllScene() as $type ) {
-					$this->innslagTyper->add( $type );
-				}
-				$this->innslagTyper->add( innslag_typer::getByName('video') );
-				$this->innslagTyper->add( innslag_typer::getByName('utstilling') );
-			}
-			if( $this->innslagTyper->getAntall() == 0 ) {
-				foreach( innslag_typer::getAllScene() as $type ) {
-					$this->innslagTyper->add( $type );
-				}
-				$this->innslagTyper->add( innslag_typer::getByName('video') );
-				$this->innslagTyper->add( innslag_typer::getByName('utstilling') );
-			}
 		}
 		return $this->innslagTyper;
 	}
