+++ write_innslag.class.php
@@ -1,18 +1,6 @@
 <?php
-/**
-	HVORDAN CONSTRUCT OG SET FUNKER:
-	- konstruktøren kjører parent::construct
-	- alle settere er allerede overskrevet fra write-klassen
-		- setter sjekker om objektet er lastet inn ($this->_loaded), noe det ikke er
-		- logger derfor endring (change) og setter verdien via parent::setter
-	- etter foreldre-konstruktøren er ferdig resetter vi changes
-		changes brukes av save-funksjonen for å avgjøre hvilke verdier som er endret
-	- alle settere sjekker om getteren gir samme verdi som setteren før den logger endring (change)
-**/	
-require_once('UKM/logger.class.php');
-require_once('UKM/innslag.class.php');
-// For valideringen.
-require_once('UKM/advarsel.class.php');
+
+require_once('UKM/log.class.php');
 
 class innslag_writeable extends innslag {
 	var $changes = array();
